k256: use `ecdsa` crate low-S normalization support
Impls the newly added `ecdsa::NormalizeLow` trait, which enables using
the `ecdsa::Signature::normalize_s` method.
